+++ plugins/auth-backend-module-gs/src/module.ts
@@ -3,11 +3,11 @@ import {
   createBackendModule,
 } from '@backstage/backend-plugin-api';
 import { providers } from '@backstage/plugin-auth-backend';
-import {
-  authProvidersExtensionPoint,
-  createOAuthProviderFactory,
-} from '@backstage/plugin-auth-node';
-import { githubAuthenticator } from '@backstage/plugin-auth-backend-module-github-provider';
+import { authProvidersExtensionPoint } from '@backstage/plugin-auth-node';
+import { createAuthProviderFactory } from './auth/createAuthProviderFactory';
+
+const OIDC_PROVIDER_NAME_PREFIX = 'oidc-';
+const CUSTOM_OIDC_PROVIDER_NAME_PREFIX = 'gs-';
 
 /** @public */
 export const authModuleGsProviders = createBackendModule({
@@ -23,24 +23,32 @@ export const authModuleGsProviders = createBackendModule({
       async init({ providersExtensionPoint, config, logger }) {
         const providersConfig = config.getConfig('auth.providers');
         const configuredProviders: string[] = providersConfig?.keys() || [];
-        const gsProviders = configuredProviders.filter(provider =>
-          provider.startsWith('gs-'),
+        const customProviders = configuredProviders.filter(
+          provider =>
+            provider.startsWith(OIDC_PROVIDER_NAME_PREFIX) ||
+            provider.startsWith(CUSTOM_OIDC_PROVIDER_NAME_PREFIX),
         );
 
-        for (const providerName of gsProviders) {
+        for (const providerName of customProviders) {
           try {
             logger.info(`Configuring auth provider: ${providerName}`);
-            const providerConfig = providersConfig
-              .getConfig(providerName)
-              .getConfig(config.getString('auth.environment'));
-            const metadataUrl = providerConfig.getString('metadataUrl');
-            const response = await fetch(new URL(metadataUrl));
-            if (!response.ok) {
-              throw new Error(response.statusText);
+
+            if (providerName.startsWith(OIDC_PROVIDER_NAME_PREFIX)) {
+              const providerConfig = providersConfig
+                .getConfig(providerName)
+                .getConfig(config.getString('auth.environment'));
+              const metadataUrl = providerConfig.getString('metadataUrl');
+              const response = await fetch(new URL(metadataUrl));
+              if (!response.ok) {
+                throw new Error(response.statusText);
+              }
             }
+
             providersExtensionPoint.registerProvider({
               providerId: providerName,
-              factory: providers.oidc.create(),
+              factory: providerName.startsWith(CUSTOM_OIDC_PROVIDER_NAME_PREFIX)
+                ? createAuthProviderFactory()
+                : providers.oidc.create(),
             });
           } catch (err) {
             logger.error(
@@ -49,31 +57,6 @@ export const authModuleGsProviders = createBackendModule({
             logger.error((err as Error).toString());
           }
         }
-
-        /**
-         * Custom sign-in resolver for GitHub auth provider that looks up the user
-         * by matching their GitHub username to the entity metadata.name
-         */
-        providersExtensionPoint.registerProvider({
-          providerId: 'github',
-          factory: createOAuthProviderFactory({
-            authenticator: githubAuthenticator,
-            async signInResolver(info, ctx) {
-              const { fullProfile } = info.result;
-
-              const userId = fullProfile.username;
-              if (!userId) {
-                throw new Error(
-                  `GitHub user profile does not contain a username`,
-                );
-              }
-
-              return ctx.signInWithCatalogUser({
-                filter: { 'metadata.name': userId },
-              });
-            },
-          }),
-        });
       },
     });
   },