+++ plugins/gs/src/apis/kubernetes-auth-providers/types.ts
@@ -1,0 +1,14 @@
+import { createApiRef } from '@backstage/core-plugin-api';
+
+export interface KubernetesAuthProvider {
+  getCredentials(): Promise<{ token?: string }>;
+}
+
+export const gsKubernetesAuthProvidersApiRef =
+  createApiRef<KubernetesAuthProvidersApi>({
+    id: 'plugin.gs.kubernetes-auth-providers',
+  });
+
+export interface KubernetesAuthProvidersApi {
+  getCredentials(authProvider: string): Promise<{ token?: string }>;
+}