+++ plugins/gs/src/components/hooks/useClusters.ts
@@ -9,11 +9,11 @@ import {
 import { useOrganizations } from './useOrganizations';
 import { useInstallations } from './useInstallations';
 import { useApi } from '@backstage/core-plugin-api';
-import { kubernetesApiRef } from '@backstage/plugin-kubernetes';
 import { useQueries } from '@tanstack/react-query';
 import { getK8sListPath } from './utils/k8sPath';
 import { getInstallationsQueriesInfo } from './utils/queries';
 import { useApiVersionOverrides } from './useApiVersionOverrides';
+import { gsKubernetesApiRef } from '../../apis/kubernetes';
 
 function getInstallationOrganizationNamespaces(
   installationName: string,
@@ -47,7 +47,7 @@ export function useClusters(installations?: string[]) {
   const { resources: organizations, initialLoading: isLoadingOrganizations } =
     useOrganizations(selectedInstallations);
 
-  const kubernetesApi = useApi(kubernetesApiRef);
+  const kubernetesApi = useApi(gsKubernetesApiRef);
 
   const queries = useQueries({
     queries: selectedInstallations.map(installationName => {